Extract `Instantiable`s into proper classes with `Cls` suffix
- Add comment for union type aliases we rewrite to inheritance, document in readme
- Tweak cleanup script so it doesn't clean sbt plugin
- Add a few libraries
- Fix a bug where contents of stdlib ends up in libraries
- Tweak order of typescript phases, add more caching
- Fix bugs with type lookups
- Fix a performance issue with `ParentsResolver`
- Fix bug in `RemoveMultipleInheritance`
- cache hashcode for trees

+++ importer/src/main/scala/com/olvind/tso/importer/BintrayCleanup.scala
@@ -31,13 +31,15 @@ object BintrayCleanup extends App {
     repo.packages(pos)().foreach { packages =>
       packages.foreach {
         case PackageSummary(name, _) =>
-          val Pkg = repo.get(name)
-          Pkg().foreach { pkg =>
-            pkg.versions.drop(5).foreach { d =>
-              Pkg.version(d).delete().foreach {
-                case Message(msg) =>
-                  deleted.increment()
-                  println(s"Deleted $name $d $msg (${deleted.get})")
+          if (!name.startsWith("sbt-")) { // leave old versions of plugin
+            val Pkg = repo.get(name)
+            Pkg().foreach { pkg =>
+              pkg.versions.drop(5).foreach { d =>
+                Pkg.version(d).delete().foreach {
+                  case Message(msg) =>
+                    deleted.increment()
+                    println(s"Deleted $name $d $msg (${deleted.get})")
+                }
               }
             }
           }
